Moved schedule.py contents to models.py @2Cubed ... FIX ALL TEH THINGS NAOW!
